Code for 130
Also we need to make it so that when you click the view all button, it will navigate you to the subcategory page for that category. The thing is that we need the subcategory page to be dynamic also, so that it can change the items displayed depending on which category you accessed the page from.